# doc/api-documentation.yml
openapi: 3.1.0
info:
  title: Connectly API
  description: |-
    Backend for a social platform: users, posts, comments and likes, with JWT authentication, email verification, password reset (link or OTP) and soft account deletion with a 30-day grace period.

    Send `Authorization: Bearer <token>` on every endpoint outside `/auth/**`. Obtain a token from `POST /auth/login`.
  contact:
    name: jomariabejo
    url: https://github.com/jomariabejo/connectly-api
  license:
    name: MIT
    url: https://github.com/jomariabejo/connectly-api/blob/main/LICENSE
  version: 0.0.1-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Local development
...
paths:
...
  /auth/reset-password:
    post:
      tags:
      - Authentication
      summary: Complete a password reset
      description: "Supply either the emailed token or the OTP, together with the\
        \ new password. Tokens expire after security.password.reset.expiry.minutes\
        \ and are single-use."
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
        required: true
      responses:
        "200":
          description: Password changed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GenericResponseString"
        "400":
          description: "Token invalid, expired, already used, or the new password\
            \ is too weak"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GenericResponseString"
        "500":
          description: Unexpected failure while resetting
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GenericResponseString"
  /auth/registration:
    post:
      tags:
      - Authentication
      summary: Register a new account
      description: "Creates a disabled account and publishes OnRegistrationCompleteEvent,\
        \ which mails a verification token valid for 24 hours. The account cannot\
        \ log in until it is verified."
      operationId: registerUserAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserDto"
        required: true
      responses:
        "200":
          description: Account created; verification email dispatched
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        "400":
          description: Validation failed on the request body
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        "409":
          description: Username or email already registered
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDto"
  /auth/login:
    post:
      tags:
      - Authentication
      summary: Log in and obtain a JWT
      description: "Returns a bearer token plus its lifetime in milliseconds (security.jwt.expiration,\
        \ 1h by default). Send it as `Authorization: Bearer <token>` on every non-/auth\
        \ endpoint."
      operationId: authenticate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginUserDto"
        required: true
      responses:
        "200":
          description: Authenticated; token returned
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: "Bad credentials, or the account is not yet verified"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /auth/forgot-password/otp:
    post:
      tags:
      - Authentication
      summary: Request a password reset one-time code
      description: "Same neutral response and rate limiting as the email variant,\
        \ but mails a short OTP instead of a link. The OTP allows security.password.reset.max-otp-attempts\
        \ tries before it is burned."
      operationId: forgotPasswordOtp
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
        required: true
      responses:
        "200":
          description: Request accepted (sent only if the account exists)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GenericResponseString"
        "400":
          description: Malformed email address
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GenericResponseString"
        "429":
          description: Too many reset attempts for this address
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GenericResponseString"
  /auth/forgot-password/email:
    post:
      tags:
      - Authentication
      summary: Request a password reset link by email
      description: Always answers with the same neutral message so the endpoint cannot
        be used to enumerate registered addresses. Rate limited by RateLimitingService.
      operationId: forgotPasswordEmail
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
        required: true
      responses:
        "200":
          description: Request accepted (sent only if the account exists)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GenericResponseString"
        "400":
          description: Malformed email address
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GenericResponseString"
        "429":
          description: Too many reset attempts for this address
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GenericResponseString"
...
  /users/me:
    get:
      tags:
      - Users
      summary: Get the caller's profile
      operationId: authenticatedUser
      responses:
        "200":
          description: Profile returned
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        "401":
          description: Missing or invalid bearer token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDto"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Users
      summary: Delete the caller's account (soft delete)
      description: "Marks the account deleted and schedules permanent removal 30 days\
        \ out, then mails a reactivation token. Returns 202 Accepted, not 204 -- deletion\
        \ is scheduled, not immediate. The body carries both timestamps and the grace\
        \ period in days."
      operationId: deleteAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountRequestDto"
      responses:
        "202":
          description: Deletion scheduled; reactivation token issued
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DeleteAccountResponseDto"
        "401":
          description: Missing or invalid bearer token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DeleteAccountResponseDto"
      security:
      - bearerAuth: []
...
  /auth/verify:
    get:
      tags:
      - Authentication
      summary: Verify an email address
      description: "Validates the token and its expiry date, then enables the account."
      operationId: verifyAccount
      parameters:
      - name: token
        in: query
        description: Verification token from the registration email
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Email verified
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Invalid or expired verification token
          content:
            '*/*':
              schema:
                type: string
  /auth/registrationConfirm:
    get:
      tags:
      - Authentication
      summary: Confirm registration from the emailed link
      description: Enables the account and consumes the verification token. This is
        the link target used by the registration email; see /auth/verify for the equivalent
        that validates expiry.
      operationId: confirmRegistration
      parameters:
      - name: token
        in: query
        description: Verification token from the registration email
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Account activated
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Unknown verification token
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
...
    UserResponseDto:
      type: object
      description: A user profile. Never carries credentials or verification tokens.
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: helloworld
        email:
          type: string
          example: test@gmail.com
        firstName:
          type: string
          example: Jomari
        lastName:
          type: string
          example: Abejo
        enabled:
          type: boolean
          description: False until the email address has been verified
          example: true
        roles:
          type: array
          description: "Granted role names, without the ROLE_ prefix"
          example:
          - USER
          items:
            type: string
          uniqueItems: true
        deletedAt:
          type: string
          format: date-time
          description: Null unless the account is scheduled for deletion
        scheduledDeletionAt:
          type: string
          format: date-time
          description: When a scheduled account is permanently removed; null otherwise
        autoReactivationEnabled:
          type: boolean
          description: Whether logging in during the grace period restores the account
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ResetPasswordRequest:
      type: object
      description: "Completes a password reset. Supply either token (email link flow)\
        \ or otp (code flow), not both."
      properties:
        token:
          type: string
          description: Token from the reset email; leave null when using an OTP
          example: your-reset-token-here
        otp:
          type: string
          description: One-time code from the reset email; leave null when using a
            token
          example: 123456
        newPassword:
          type: string
          description: "New password, at least security.password.validation.min-length\
            \ characters"
          example: NewSecurePassword123!
    GenericResponseString:
      type: object
      properties:
        message:
          type: string
        data:
          type: string
    RegisterUserDto:
      type: object
      description: New account details. The account starts disabled until the emailed
        verification token is used.
      properties:
        email:
          type: string
          description: Must be unique across accounts
          example: test@gmail.com
          minLength: 1
        password:
          type: string
          description: Stored BCrypt-hashed
          example: admin123
          minLength: 1
        username:
          type: string
          description: Must be unique across accounts
          example: helloworld
          minLength: 1
      required:
      - email
      - password
      - username
    LoginUserDto:
      type: object
      description: "Login credentials. Authentication is by email, not username."
      properties:
        email:
          type: string
          example: test@gmail.com
          minLength: 1
        password:
          type: string
          example: admin123
          minLength: 1
      required:
      - email
      - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        expiresIn:
          type: integer
          format: int64
    ForgotPasswordRequest:
      type: object
      description: Address to send a reset link or OTP to.
      properties:
        email:
          type: string
          description: The response is identical whether or not this address is registered
          example: user@example.com
...
    DeleteAccountRequestDto:
      type: object
      properties:
        reason:
          type: string
        autoReactivationEnabled:
          type: boolean
    DeleteAccountResponseDto:
      type: object
      properties:
        message:
          type: string
        deletedAt:
          type: string
          format: date-time
        scheduledDeletionAt:
          type: string
          format: date-time
        gracePeriodDays:
          type: integer
          format: int32
        autoReactivationEnabled:
          type: boolean
        reactivationInstructions:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT obtained from POST /auth/login. Paste the raw token; the Bearer
        prefix is added for you.
      in: header
      scheme: bearer
      bearerFormat: JWT